apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
PyJWT[crypto]==2.10.1
email-validator==2.2.0
httpx==0.28.1
pytest==8.3.4
celery[redis]==5.4.0
